Make all hsStream reads and writes size-explicit.
This also makes the placement readers and writers type-safe.  However,
unfortunately the non-placement readers cannot be made type-safe since
they return an element rather than accepting one as a parameter.  For
now, both styles are kept since both have their utility independently
of each other.

+++ Sources/Plasma/PubUtilLib/plDrawable/plSpaceTree.cpp
@@ -76,7 +76,7 @@ void plSpaceTreeNode::Write(hsStream* s)
     s->WriteLE16(fChildren[0]);
     if( fFlags & kIsLeaf )
         // Temp for now to play nice with binary patches
-        s->WriteLE16( 0 );
+        s->WriteLE16(uint16_t(0));
     else
         s->WriteLE16(fChildren[1]);
 }
